+++ packages/expo-notifications/android/src/main/java/expo/modules/notifications/service/delegates/ExpoPresentationDelegate.kt
@@ -12,6 +12,7 @@ import android.service.notification.StatusBarNotification
 import android.util.Log
 import android.util.Pair
 import androidx.core.app.NotificationManagerCompat
+import expo.modules.notifications.notifications.SoundResolver
 import expo.modules.notifications.notifications.enums.NotificationPriority
 import expo.modules.notifications.notifications.model.Notification
 import expo.modules.notifications.notifications.model.NotificationBehavior
@@ -20,6 +21,9 @@ import expo.modules.notifications.notifications.model.NotificationRequest
 import expo.modules.notifications.notifications.presentation.builders.CategoryAwareNotificationBuilder
 import expo.modules.notifications.notifications.presentation.builders.ExpoNotificationBuilder
 import expo.modules.notifications.service.interfaces.PresentationDelegate
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import org.json.JSONException
 import org.json.JSONObject
 import java.util.Date
@@ -88,7 +92,7 @@ open class ExpoPresentationDelegate(
    * set badge count without showing a notification.
    */
   override fun presentNotification(notification: Notification, behavior: NotificationBehavior?) {
-    if (behavior != null && !behavior.shouldShowAlert()) {
+    if (behavior?.shouldShowAlert() == false) {
       if (behavior.shouldPlaySound()) {
         val sound = getNotificationSoundUri(notification) ?: Settings.System.DEFAULT_NOTIFICATION_URI
         RingtoneManager.getRingtone(
@@ -98,11 +102,18 @@ open class ExpoPresentationDelegate(
       }
       return
     }
-    NotificationManagerCompat.from(context).notify(
-      notification.notificationRequest.identifier,
-      getNotifyId(notification.notificationRequest),
-      createNotification(notification, behavior)
-    )
+    CoroutineScope(Dispatchers.IO).launch {
+      val androidNotification = CategoryAwareNotificationBuilder(context, SharedPreferencesNotificationCategoriesStore(context)).apply {
+        setNotification(notification)
+        setAllowedBehavior(behavior)
+      }.build()
+
+      NotificationManagerCompat.from(context).notify(
+        notification.notificationRequest.identifier,
+        getNotifyId(notification.notificationRequest),
+        androidNotification
+      )
+    }
   }
 
   private fun getNotificationSoundUri(notification: Notification): Uri? {
@@ -111,7 +122,8 @@ open class ExpoPresentationDelegate(
         notificationManager.getNotificationChannel(it)?.sound
       }
     } else {
-      notification.notificationRequest.content.sound
+      val name = notification.notificationRequest.content.soundName
+      SoundResolver(context).resolve(name)
     }
   }
 
@@ -147,15 +159,9 @@ open class ExpoPresentationDelegate(
 
   override fun dismissAllNotifications() = NotificationManagerCompat.from(context).cancelAll()
 
-  protected open fun createNotification(notification: Notification, notificationBehavior: NotificationBehavior?): android.app.Notification =
-    CategoryAwareNotificationBuilder(context, SharedPreferencesNotificationCategoriesStore(context)).also {
-      it.setNotification(notification)
-      it.setAllowedBehavior(notificationBehavior)
-    }.build()
-
   protected open fun getNotification(statusBarNotification: StatusBarNotification): Notification? {
     val notification = statusBarNotification.notification
-    notification.extras.getByteArray(ExpoNotificationBuilder.EXTRAS_MARSHALLED_NOTIFICATION_REQUEST_KEY)?.let {
+    notification.extras.getByteArray(ExpoNotificationBuilder.Companion.EXTRAS_MARSHALLED_NOTIFICATION_REQUEST_KEY)?.let {
       try {
         with(Parcel.obtain()) {
           this.unmarshall(it, 0, it.size)